Zonemaster::LDNS::RR: NSEC3 RRs never cover root
By definition, no NSEC3 resource record can ever prove the nonexistence
of the root zone. However, calling the covers() method on an object
representing an NSEC3 resource record, with "." as input, caused a
crash. This commit changes the function to return undef in this case,
which can be interpreted as a false value.