Zonemaster::LDNS::RR::NSEC3: rewrite next_owner()
The next_owner() method in Zonemaster::LDNS::RR::NSEC3 did not work as
documented.

Its return value was not the value of the next hashed owner name field,
but the same value with the length byte prepended. This choice does not
make the function as useful as one may hope.

This commit ensures that the next_owner() method returns the next hashed
owner name field, and only that.

The unit test for this function compares base64-encoded strings. This
hampers legibility because conversions between base32 and base64 are not
trivial and only a very astute reader would notice that the next_owner()
method had this kind of catch.

As a bonus, add a useful tip in the method’s documentation.