Zonemaster::LDNS::RR::NSEC3PARAM: rewrite salt()
The salt() method in Zonemaster::LDNS::RR::NSEC3PARAM did not work as
documented.

Like two of those accessor methods in NSEC3 I’ve fixed previously, its
return value is a byte string that starts with an extraneous length
byte, which isn’t necessary.

The unit test for this function compares base64-encoded strings. This
hampers legibility because conversions between hexadecimal strings and
base64 strings are not trivial and only a very astute reader would
notice that the salt() method also had a similar problem.

Also improve the documentation and the unit test coverage for that
method.