Fix compiler warning in XS code
The return value of ldns_rr_descript(), of type const
ldns_rr_descriptor*, was assigned to a variable of type
ldns_rr_descriptor*, thus dropping the const qualifier.

Making said variable a const pointer fixes the warning. And it’s safe to
do so too, because no modifications are made through that pointer.