to_idn(): croak with locale-insensitive message
The to_idn() function in Zonemaster::LDNS croaks with a plain string if
an error condition is encountered when converting the supplied string to
an IDN A-label.

The test suite provokes one of those error conditions in order to check
that a string is indeed thrown. That string is matched against a regular
expression.

But when constructing the string, idn2_strerror() is used, which
produces a human-readable explanation of the error but in a
locale-dependent way. So the test can succeed on one developer’s machine
and fail on someone else’s whose system language is different.

However, there is little point in replacing these plain strings with
exception objects, such as a hypothetical Zonemaster::LDNS::Error class
hierarchy. The only other place in the codebase where to_idn() is called
is in one file in Zonemaster::Engine, and it doesn’t even care what
specific error happened during IDN conversion. Exception objects would
work, but they are a pain to build in XS code (even calling out to a
constructor function written in Perl is extremely verbose in XS) and
would be an extremely overengineered solution to our problem.

Fortunately, instead of idn2_strerror(), we can use
idn2_strerror_name(), which returns locale-independent strings based on
the names of the source code’s macros. This solution is very simple,
makes the error messages predictable and does not hamper debuggability
of Zonemaster::LDNS::to_idn() too much.

Of course, should Zonemaster::Engine want to display what specifically
is wrong with the domain name they entered, we might eventually need
the alternative solution.