rr.t: No need for fc() anymore
Remove a case-insensitive comparison that was done in t/rr.t in order to
make unit tests pass on installations where the now removed case
randomization feature was enabled.

Previously, this equality comparison was done case-sensitively, which
only worked when this feature was disabled.

With this change, we can now be almost certain that we don’t do any case
alterations anymore.