Allow construction of empty RRLists
The change in #203 introduced a new() function in the
Zonemaster::LDNS::RRList module.

However, calling new() with a ref to an empty array as paramater caused
the code to croak with the error message "List is empty". However, there
are situations where it may be desirable to construct such an empty
RRList.

This commit ensures that Zonemaster::LDNS::RRList->new([]) works as
intended.