Makefile.PL: play nicely with Devel::CheckLib
When pkg-config is used to discover the appropriate header and library
locations, the new code forgot to update a hash that is passed to the
cc_assert_lib function later on.

This means that on systems with non-traditional locations for openssl or
ldns libraries, the cc_assert_lib didn’t check the locations previously
discovered by pkg-config.

This commit should ensure that cc_assert_lib finds those libraries when
testing for their presence.