Improve performance by not “remembering” objects
This commit removes code that has had a very long history and legacy. It
survived over 10 years in the code base without much documentation of
its purpose, either in code comments, documentation or commit messages.

It’s likely that it might have been part of some debugging code paths
that allow a programmer to retrieve access to objects touched by LDNS.
It stores all of these objects within a few hashes.

However, maintaining these hashes proves to be rather expensive.
Disabling it speeds up simple runs of zonemaster-cli by roughly 3%,
which is significant when Zonemaster is used to test millions of
domains.