Update obsolete unit test
Running make test with TEST_WITH_NETWORK=1 in the environment fails
because the mail exchange (MX) records for iis.se were changed. Now,
it’s a single record of preference 0 pointing to something within the
“microsoft” TLD.

I tried to retain the spirit of the original test by picking another
domain with 2 MX records with non-zero preference values and whose MX
records have a reasonable expectation of stability.